chore: ajuste de pastas no frontend com divisão de responsabilidades entre componentes
