--- Frontend/esboco/style.css
@@ -1,143 +1,0 @@
-* {
-  padding: 0;
-  margin: 0;
-}
-
-:root {
-  --color1: #0087e1;
-  --color2: #e05d00;
-  --color3: #e09f00;
-  --color4: #00f543;
-  --color5: #f50400;
-}
-
-nav {
-  display: flex;
-  text-align: center;
-  justify-content: space-around;
-  background-color: var(--color1);
-  color: #ffffff;
-}
-
-nav ul {
-  margin: 0 auto;
-  width: 50%;
-  display: flex;
-  padding: 1em;
-  justify-content: space-between;
-  list-style: none;
-}
-
-nav li {
-  margin: 1em;
-}
-
-nav a {
-  color: #ffffff;
-  text-decoration: none;
-}
-
-nav a:hover {
-  color: var(--color2);
-}
-
-table {
-  width: 70%;
-  border-collapse: collapse;
-  margin: 20px auto;
-}
-
-th,
-td {
-  border: 1px solid #999;
-  padding: 8px 12px;
-  text-align: center;
-}
-
-th {
-  background-color: var(--color1);
-  color: #ffffff;
-}
-
-th button {
-  background-color: var(--color2);
-  border: none;
-  color: #ffffff;
-  border-radius: 0.5em;
-  padding: 0.2em;
-}
-
-th button:hover {
-  cursor: pointer;
-  opacity: 0.7;
-}
-
-tr button {
-  border: none;
-  color: #ffffff;
-  border-radius: 0.5em;
-  padding: 0.2em;
-  margin: 0.5em;
-}
-tr button:hover {
-  cursor: pointer;
-  opacity: 0.8;
-}
-
-tr button:first-child {
-  background-color: var(--color4);
-}
-
-tr button:nth-child(2) {
-  background-color: var(--color3);
-}
-
-tr button:nth-child(3) {
-  background-color: var(--color5);
-}
-
-h1 {
-  color: var(--color1);
-  text-align: center;
-}
-form {
- 
-  display: flex;
-  flex-direction: column;
-  margin: 0 auto;
-  width: 50%;
-  padding: .5em;
-}
-form label {
-  margin: 0 auto;
-  color: var(--color1);
-  text-align: start;
-  margin-top: .5em;
-  margin-bottom: .5em;
-}
-form input,select,option {
-  margin: 0 auto;
-  
-  width: 40%;
-}
-form button {
-  background-color: var(--color2);
-  border: none;
-  color: #ffffff;
-  border-radius: 0.5em;
-  margin: 0 auto;
-  width: 25%;
-  margin-top: 1em;
-}
-form button:hover {
- cursor: pointer;
- opacity: 0.8;
-}
-#deleteProduct
-{
-  background-color: #f50400;
-}
-
-footer {
-  text-align: center;
-}
