+++ Frontend/inventorySystem/src/App.css
@@ -98,12 +98,11 @@ h1 {
   text-align: center;
 }
 form {
- 
   display: flex;
   flex-direction: column;
   margin: 0 auto;
-  width: 50%;
-  padding: .5em;
+  width: 80%;
+  padding: 5em;
 }
 form label {
   margin: 0 auto;
