--- Frontend/inventorySystem/src/Components/Main.tsx
@@ -1,14 +1,0 @@
-import TableOperations from "./TableOperations"
-import TableProducts from "./TableProducts"
-
-
-const Main = () => {
-  return (
-    <main>
-     
-    <TableOperations/>
-    </main>
-  )
-}
-
-export default Main
