+++ Frontend/inventorySystem/src/Components/form/AddOperationForm.tsx
