+++ Frontend/inventorySystem/src/Components/layout/NavBar.tsx
